Move utility classes to utilities directory
