+++ src/main/java/frc/robot/Constants.java
@@ -21,6 +21,8 @@ import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
 import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.math.util.Units;
+import frc.robot.utilities.PIDF;
+import frc.robot.utilities.TrapezoidalConstraint;
 
 public final class Constants {
   public static final double kDt = 0.02;
