+++ src/main/java/frc/robot/commands/JoystickTargetNote.java
@@ -9,10 +9,10 @@ import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.NoteConstants;
-import frc.robot.PIDF;
-import frc.robot.TunablePIDF;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.Limelight;
+import frc.robot.utilities.PIDF;
+import frc.robot.utilities.TunablePIDF;
 
 public class JoystickTargetNote extends Command {
   private static final TunablePIDF targetTurningPIDF =
