+++ src/main/java/frc/robot/utilities/ShooterInterp.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.utilities;
 
 import java.util.Arrays;
