+++ src/main/java/frc/robot/utilities/TunableDouble.java
@@ -1,8 +1,9 @@
-package frc.robot;
+package frc.robot.utilities;
 
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
+import frc.robot.Constants;
 
 public class TunableDouble {
   private static NetworkTable table = NetworkTableInstance.getDefault().getTable("tuning");
